PageRank use float

+++ apps/PageRank.C
@@ -26,9 +26,9 @@
 
 template <class vertex>
 struct PR_F {
-  double* p_curr, *p_next;
+  float* p_curr, *p_next;
   vertex* V;
-  PR_F(double* _p_curr, double* _p_next, vertex* _V) : 
+  PR_F(float* _p_curr, float* _p_next, vertex* _V) : 
     p_curr(_p_curr), p_next(_p_next), V(_V) {}
   inline bool update(uintE s, uintE d){ //update function applies PageRank equation
     p_next[d] += p_curr[s]/V[s].getOutDegree();
@@ -43,14 +43,14 @@ struct PR_F {
 //vertex map function to update its p value according to PageRank equation
 template <class vertex>
 struct PR_Vertex_F {
-  double damping;
-  double addedConstant;
-  double* p_curr;
-  double* p_next;
+  float damping;
+  float addedConstant;
+  float* p_curr;
+  float* p_next;
   vertex* V;
-  PR_Vertex_F(double* _p_curr, double* _p_next, double _damping, intE n, vertex* _V) :
+  PR_Vertex_F(float* _p_curr, float* _p_next, float _damping, intE n, vertex* _V) :
     p_curr(_p_curr), p_next(_p_next), 
-    damping(_damping), addedConstant((1-_damping)*(1/(double)n)), V(_V){}
+    damping(_damping), addedConstant((1-_damping)*(1/(float)n)), V(_V){}
   inline bool operator () (uintE i) {
     p_next[i] = damping*p_next[i] + addedConstant;
     if (V[i].getOutDegree() != 0) p_next[i] /= V[i].getOutDegree();
@@ -60,8 +60,8 @@ struct PR_Vertex_F {
 
 //resets p
 struct PR_Vertex_Reset {
-  double* p_curr;
-  PR_Vertex_Reset(double* _p_curr) :
+  float* p_curr;
+  PR_Vertex_Reset(float* _p_curr) :
     p_curr(_p_curr) {}
   inline bool operator () (uintE i) {
     p_curr[i] = 0.0;
@@ -73,12 +73,12 @@ template <class vertex>
 void Compute(graph<vertex>& GA, commandLine P) {
   long maxIters = P.getOptionLongValue("-maxiters",100);
   const intE n = GA.n;
-  const double damping = 0.85, epsilon = 0.0000001;
+  const float damping = 0.85, epsilon = 0.0000001;
   
-  double one_over_n = 1/(double)n;
-  double* p_curr = newA(double,n);
+  float one_over_n = 1/(float)n;
+  float* p_curr = newA(float,n);
   {parallel_for(long i=0;i<n;i++) p_curr[i] = one_over_n;}
-  double* p_next = newA(double,n);
+  float* p_next = newA(float,n);
   {parallel_for(long i=0;i<n;i++) p_next[i] = 0;} //0 if unchanged
   bool* frontier = newA(bool,n);
   {parallel_for(long i=0;i<n;i++) frontier[i] = 1;}
@@ -93,7 +93,7 @@ void Compute(graph<vertex>& GA, commandLine P) {
     {parallel_for(long i=0;i<n;i++) {
       p_curr[i] = fabs(p_curr[i]-p_next[i]);
       }}
-    double L1_norm = sequence::plusReduce(p_curr,n);
+    float L1_norm = sequence::plusReduce(p_curr,n);
     if(L1_norm < epsilon) break;
     //reset p_curr
     vertexMap(Frontier,PR_Vertex_Reset(p_curr));
